Create Chart class that all charts inherit common properties and methods from.
